isolate parseField into its own file
